requirements.txt
asgiref==3.3.1
blessings==1.7
bpython==0.21
certifi==2020.12.5
chardet==4.0.0
curtsies==0.3.5
cwcwidth==0.1.2
Django==3.1.6
django-ckeditor==6.0.0
django-extensions==3.1.0
django-js-asset==1.2.2
django-redis==4.12.1
django-redis-cache==3.0.0
greenlet==1.0.0
idna==2.10
mysqlclient==2.0.3
Pygments==2.7.4
python-decouple==3.4
pytz==2021.1
pyxdg==0.27
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
